+++ packages/project/src/serialize/index.ts
@@ -1,5 +1,5 @@
-import json from './to-json';
 import state from './to-app-state';
 import fs from './to-fs';
+import project from './to-project';
 
-export { json, state, fs };
+export { project, state, fs };