+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @openfn/compiler
 
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [72f18b1]
+- Updated dependencies [2cc28a6]
+  - @openfn/lexicon@1.2.7
+  - @openfn/logger@1.1.0
+
 ## 1.2.0
 
 ### Minor Changes