+++ packages/ws-worker/CHANGELOG.md
@@ -1,5 +1,16 @@
 # ws-worker
 
+## 1.19.7
+
+### Patch Changes
+
+- Updated dependencies [72f18b1]
+- Updated dependencies [2cc28a6]
+  - @openfn/lexicon@1.2.7
+  - @openfn/logger@1.1.0
+  - @openfn/engine-multi@1.8.5
+  - @openfn/runtime@1.7.6
+
 ## 1.19.6
 
 ### Patch Changes